% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusindia-package.R
\docType{package}
\name{censusindia-package}
\alias{censusindia}
\alias{censusindia-package}
\title{censusindia: Census of India data for R}
\description{
Digitised Census of India data from 1901 to 2011: population time series, primary census abstracts at state and district levels, mother tongue tables, SC/ST and tribal populations, and administrative directories. Census variables can be retrieved with geographic boundaries attached for mapping.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://saketlab.github.io/censusindia/}
  \item \url{https://github.com/saketlab/censusindia}
  \item Report bugs at \url{https://github.com/saketlab/censusindia/issues}
}

}
\author{
\strong{Maintainer}: Saket Choudhary \email{saketkc@gmail.com}

Authors:
\itemize{
  \item Saket Choudhary \email{saketkc@gmail.com}
}

}
\keyword{internal}
